--- a/pkg/config/host/names.go
+++ b/pkg/config/host/names.go
@@ -32,7 +32,10 @@ func (h Names) Len() int {
 }
 
 func (h Names) Less(i, j int) bool {
-	a, b := h[i], h[j]
+	return MoreSpecific(h[i], h[j])
+}
+
+func MoreSpecific(a, b Name) bool {
 	if len(a) == 0 && len(b) == 0 {
 		return true // doesn't matter, they're both the empty string
 	}