--- a/pkg/config/host/names.go
+++ b/pkg/config/host/names.go
@@ -35,6 +35,7 @@ func (h Names) Less(i, j int) bool {
 	return MoreSpecific(h[i], h[j])
 }
 
+// MoreSpecific returns true if hostname a is more specific than b.
 func MoreSpecific(a, b Name) bool {
 	if len(a) == 0 && len(b) == 0 {
 		return true // doesn't matter, they're both the empty string