--- a/releasenotes/notes/optimize-most-specific-host-match.yaml
+++ b/releasenotes/notes/optimize-most-specific-host-match.yaml
@@ -1,0 +1,6 @@
+apiVersion: release-notes/v2
+kind: feature
+area: traffic-management
+releaseNotes:
+- |
+  **Optimized** MostSpecificHostMatch to prevent fullscanning hosts when encountering wildcards.