**Please provide a description of this PR:**
Attempt to improve #41453 and #41454 :
- Split handling of specific and wildcard domains to improve worst-case scenario matches on mixed environments.
- Modify MostSpecificHostMatch to also return the value of the match, preventing additional lookups
- Optimize MostSpecificHostMatch to not allocate a slice of `host.Name` for sorting

This changes should show no degradation of performance for most scenarios, but show improvements for scenarios of mixed wildcard and specific destinationrules.
It would also show some improvements whenever multiple destinationrules match and the longest matching is chosen.