@howardjohn 2022-10-24
Will take a look soon - do we have any benchmark numbers? maybe even running the one from the other PR