@sschepens 2022-10-24
@howardjohn there is an included benchmark but it doesn't cover multiple matching destination rules or mixed wildcard+specific use cases which should theoretically be common. I'll see to modifying these benchmarks.

I'll also see about the Benchmark in the other PR.