@sschepens 2022-10-24
with the benchmarks in #41614 :
Weirdly `benchstat` doesn't recognize as there being any improvements, but you can certainly see the time and memory difference, might be related to the function being very small and quick.
New version does no allocations and provides speedups in various cases with no big regression.
```
name                                                      old time/op    new time/op    delta
MostSpecificHostMatch/10ExactNoMatch-8                       107ns ± 0%      13ns ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/50ExactNoMatch-8                       424ns ± 0%      13ns ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/100ExactNoMatch-8                      848ns ± 0%      13ns ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/1000ExactNoMatch-8                    9.62µs ± 0%    0.01µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/5000ExactNoMatch-8                    44.6µs ± 0%     0.0µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/10ExactMatch-8                        10.2ns ± 0%     9.6ns ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/50ExactMatch-8                        11.1ns ± 0%    11.8ns ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/100ExactMatch-8                       10.3ns ± 0%    10.5ns ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/1000ExactMatch-8                      8.96ns ± 0%    8.94ns ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/5000ExactMatch-8                      10.4ns ± 0%    10.1ns ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/10DestRuleWildcardNoMatch-8            143ns ± 0%     128ns ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/50DestRuleWildcardNoMatch-8            472ns ± 0%     478ns ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/100DestRuleWildcardNoMatch-8          1.22µs ± 0%    1.13µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/1000DestRuleWildcardNoMatch-8         13.0µs ± 0%    12.8µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/5000DestRuleWildcardNoMatch-8         55.2µs ± 0%    55.4µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/10DestRuleWildcardMatch-8              170ns ± 0%     156ns ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/50DestRuleWildcardMatch-8              642ns ± 0%     641ns ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/100DestRuleWildcardMatch-8            1.25µs ± 0%    1.19µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/1000DestRuleWildcardMatch-8           12.9µs ± 0%    12.8µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/5000DestRuleWildcardMatch-8           66.7µs ± 0%    66.6µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/10NeedleWildcardNoMatch-8              156ns ± 0%     140ns ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/50NeedleWildcardNoMatch-8              617ns ± 0%     609ns ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/100NeedleWildcardNoMatch-8            1.25µs ± 0%    1.23µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/1000NeedleWildcardNoMatch-8           13.5µs ± 0%    13.1µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/5000NeedleWildcardNoMatch-8           52.6µs ± 0%    53.3µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/10NeedleWildcardMatch-8                178ns ± 0%     157ns ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/50NeedleWildcardMatch-8                698ns ± 0%     639ns ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/100NeedleWildcardMatch-8              1.29µs ± 0%    1.17µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/1000NeedleWildcardMatch-8             13.7µs ± 0%    13.2µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatch/5000NeedleWildcardMatch-8             70.7µs ± 0%    68.2µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/10DestRuleWildcardNoMatch-8       229ns ± 0%     136ns ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/50DestRuleWildcardNoMatch-8       886ns ± 0%     476ns ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/100DestRuleWildcardNoMatch-8     2.26µs ± 0%    1.16µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/1000DestRuleWildcardNoMatch-8    32.6µs ± 0%    12.9µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/5000DestRuleWildcardNoMatch-8     165µs ± 0%      55µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/10DestRuleWildcardMatch-8         254ns ± 0%     152ns ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/50DestRuleWildcardMatch-8        1.12µs ± 0%    0.62µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/100DestRuleWildcardMatch-8       2.32µs ± 0%    1.16µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/1000DestRuleWildcardMatch-8      33.0µs ± 0%    12.9µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/5000DestRuleWildcardMatch-8       177µs ± 0%      67µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/10NeedleWildcardNoMatch-8         231ns ± 0%     144ns ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/50NeedleWildcardNoMatch-8        1.11µs ± 0%    0.61µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/100NeedleWildcardNoMatch-8       2.33µs ± 0%    1.20µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/1000NeedleWildcardNoMatch-8      32.2µs ± 0%    13.0µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/5000NeedleWildcardNoMatch-8       160µs ± 0%      53µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/10NeedleWildcardMatch-8           250ns ± 0%     208ns ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/50NeedleWildcardMatch-8          1.16µs ± 0%    0.64µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/100NeedleWildcardMatch-8         2.35µs ± 0%    1.20µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/1000NeedleWildcardMatch-8        32.4µs ± 0%    13.9µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/5000NeedleWildcardMatch-8         175µs ± 0%      69µs ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatchMultiMatch/DestRuleWildcard-8           232ns ± 0%      86ns ± 0%   ~     (p=1.000 n=1+1)
MostSpecificHostMatchMultiMatch/NeedleWildcard-8             230ns ± 0%      91ns ± 0%   ~     (p=1.000 n=1+1)

name                                                      old alloc/op   new alloc/op   delta
MostSpecificHostMatch/10ExactNoMatch-8                       0.00B          0.00B        ~     (all equal)
MostSpecificHostMatch/50ExactNoMatch-8                       0.00B          0.00B        ~     (all equal)
MostSpecificHostMatch/100ExactNoMatch-8                      0.00B          0.00B        ~     (all equal)
MostSpecificHostMatch/1000ExactNoMatch-8                     0.00B          0.00B        ~     (all equal)
MostSpecificHostMatch/5000ExactNoMatch-8                     0.00B          0.00B        ~     (all equal)
MostSpecificHostMatch/10ExactMatch-8                         0.00B          0.00B        ~     (all equal)
MostSpecificHostMatch/50ExactMatch-8                         0.00B          0.00B        ~     (all equal)
MostSpecificHostMatch/100ExactMatch-8                        0.00B          0.00B        ~     (all equal)
MostSpecificHostMatch/1000ExactMatch-8                       0.00B          0.00B        ~     (all equal)
MostSpecificHostMatch/5000ExactMatch-8                       0.00B          0.00B        ~     (all equal)
MostSpecificHostMatch/10DestRuleWildcardNoMatch-8            0.00B          0.00B        ~     (all equal)
MostSpecificHostMatch/50DestRuleWildcardNoMatch-8            0.00B          0.00B        ~     (all equal)
MostSpecificHostMatch/100DestRuleWildcardNoMatch-8           0.00B          0.00B        ~     (all equal)
MostSpecificHostMatch/1000DestRuleWildcardNoMatch-8          0.00B          0.00B        ~     (all equal)
MostSpecificHostMatch/5000DestRuleWildcardNoMatch-8          0.00B          0.00B        ~     (all equal)
MostSpecificHostMatch/10DestRuleWildcardMatch-8              16.0B ± 0%      0.0B        ~     (p=1.000 n=1+1)
MostSpecificHostMatch/50DestRuleWildcardMatch-8              16.0B ± 0%      0.0B        ~     (p=1.000 n=1+1)
MostSpecificHostMatch/100DestRuleWildcardMatch-8             16.0B ± 0%      0.0B        ~     (p=1.000 n=1+1)
MostSpecificHostMatch/1000DestRuleWildcardMatch-8            16.0B ± 0%      0.0B        ~     (p=1.000 n=1+1)
MostSpecificHostMatch/5000DestRuleWildcardMatch-8            16.0B ± 0%      0.0B        ~     (p=1.000 n=1+1)
MostSpecificHostMatch/10NeedleWildcardNoMatch-8              0.00B          0.00B        ~     (all equal)
MostSpecificHostMatch/50NeedleWildcardNoMatch-8              0.00B          0.00B        ~     (all equal)
MostSpecificHostMatch/100NeedleWildcardNoMatch-8             0.00B          0.00B        ~     (all equal)
MostSpecificHostMatch/1000NeedleWildcardNoMatch-8            0.00B          0.00B        ~     (all equal)
MostSpecificHostMatch/5000NeedleWildcardNoMatch-8            0.00B          0.00B        ~     (all equal)
MostSpecificHostMatch/10NeedleWildcardMatch-8                16.0B ± 0%      0.0B        ~     (p=1.000 n=1+1)
MostSpecificHostMatch/50NeedleWildcardMatch-8                16.0B ± 0%      0.0B        ~     (p=1.000 n=1+1)
MostSpecificHostMatch/100NeedleWildcardMatch-8               16.0B ± 0%      0.0B        ~     (p=1.000 n=1+1)
MostSpecificHostMatch/1000NeedleWildcardMatch-8              16.0B ± 0%      0.0B        ~     (p=1.000 n=1+1)
MostSpecificHostMatch/5000NeedleWildcardMatch-8              16.0B ± 0%      0.0B        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/10DestRuleWildcardNoMatch-8       0.00B          0.00B        ~     (all equal)
MostSpecificHostMatchMixed/50DestRuleWildcardNoMatch-8       0.00B          0.00B        ~     (all equal)
MostSpecificHostMatchMixed/100DestRuleWildcardNoMatch-8      0.00B          0.00B        ~     (all equal)
MostSpecificHostMatchMixed/1000DestRuleWildcardNoMatch-8     0.00B          0.00B        ~     (all equal)
MostSpecificHostMatchMixed/5000DestRuleWildcardNoMatch-8     0.00B          0.00B        ~     (all equal)
MostSpecificHostMatchMixed/10DestRuleWildcardMatch-8         16.0B ± 0%      0.0B        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/50DestRuleWildcardMatch-8         16.0B ± 0%      0.0B        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/100DestRuleWildcardMatch-8        16.0B ± 0%      0.0B        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/1000DestRuleWildcardMatch-8       16.0B ± 0%      0.0B        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/5000DestRuleWildcardMatch-8       16.0B ± 0%      0.0B        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/10NeedleWildcardNoMatch-8         0.00B          0.00B        ~     (all equal)
MostSpecificHostMatchMixed/50NeedleWildcardNoMatch-8         0.00B          0.00B        ~     (all equal)
MostSpecificHostMatchMixed/100NeedleWildcardNoMatch-8        0.00B          0.00B        ~     (all equal)
MostSpecificHostMatchMixed/1000NeedleWildcardNoMatch-8       0.00B          0.00B        ~     (all equal)
MostSpecificHostMatchMixed/5000NeedleWildcardNoMatch-8       0.00B          0.00B        ~     (all equal)
MostSpecificHostMatchMixed/10NeedleWildcardMatch-8           16.0B ± 0%      0.0B        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/50NeedleWildcardMatch-8           16.0B ± 0%      0.0B        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/100NeedleWildcardMatch-8          16.0B ± 0%      0.0B        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/1000NeedleWildcardMatch-8         16.0B ± 0%      0.0B        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/5000NeedleWildcardMatch-8         16.0B ± 0%      0.0B        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMultiMatch/DestRuleWildcard-8            136B ± 0%        0B        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMultiMatch/NeedleWildcard-8              136B ± 0%        0B        ~     (p=1.000 n=1+1)

name                                                      old allocs/op  new allocs/op  delta
MostSpecificHostMatch/10ExactNoMatch-8                        0.00           0.00        ~     (all equal)
MostSpecificHostMatch/50ExactNoMatch-8                        0.00           0.00        ~     (all equal)
MostSpecificHostMatch/100ExactNoMatch-8                       0.00           0.00        ~     (all equal)
MostSpecificHostMatch/1000ExactNoMatch-8                      0.00           0.00        ~     (all equal)
MostSpecificHostMatch/5000ExactNoMatch-8                      0.00           0.00        ~     (all equal)
MostSpecificHostMatch/10ExactMatch-8                          0.00           0.00        ~     (all equal)
MostSpecificHostMatch/50ExactMatch-8                          0.00           0.00        ~     (all equal)
MostSpecificHostMatch/100ExactMatch-8                         0.00           0.00        ~     (all equal)
MostSpecificHostMatch/1000ExactMatch-8                        0.00           0.00        ~     (all equal)
MostSpecificHostMatch/5000ExactMatch-8                        0.00           0.00        ~     (all equal)
MostSpecificHostMatch/10DestRuleWildcardNoMatch-8             0.00           0.00        ~     (all equal)
MostSpecificHostMatch/50DestRuleWildcardNoMatch-8             0.00           0.00        ~     (all equal)
MostSpecificHostMatch/100DestRuleWildcardNoMatch-8            0.00           0.00        ~     (all equal)
MostSpecificHostMatch/1000DestRuleWildcardNoMatch-8           0.00           0.00        ~     (all equal)
MostSpecificHostMatch/5000DestRuleWildcardNoMatch-8           0.00           0.00        ~     (all equal)
MostSpecificHostMatch/10DestRuleWildcardMatch-8               1.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
MostSpecificHostMatch/50DestRuleWildcardMatch-8               1.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
MostSpecificHostMatch/100DestRuleWildcardMatch-8              1.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
MostSpecificHostMatch/1000DestRuleWildcardMatch-8             1.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
MostSpecificHostMatch/5000DestRuleWildcardMatch-8             1.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
MostSpecificHostMatch/10NeedleWildcardNoMatch-8               0.00           0.00        ~     (all equal)
MostSpecificHostMatch/50NeedleWildcardNoMatch-8               0.00           0.00        ~     (all equal)
MostSpecificHostMatch/100NeedleWildcardNoMatch-8              0.00           0.00        ~     (all equal)
MostSpecificHostMatch/1000NeedleWildcardNoMatch-8             0.00           0.00        ~     (all equal)
MostSpecificHostMatch/5000NeedleWildcardNoMatch-8             0.00           0.00        ~     (all equal)
MostSpecificHostMatch/10NeedleWildcardMatch-8                 1.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
MostSpecificHostMatch/50NeedleWildcardMatch-8                 1.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
MostSpecificHostMatch/100NeedleWildcardMatch-8                1.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
MostSpecificHostMatch/1000NeedleWildcardMatch-8               1.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
MostSpecificHostMatch/5000NeedleWildcardMatch-8               1.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/10DestRuleWildcardNoMatch-8        0.00           0.00        ~     (all equal)
MostSpecificHostMatchMixed/50DestRuleWildcardNoMatch-8        0.00           0.00        ~     (all equal)
MostSpecificHostMatchMixed/100DestRuleWildcardNoMatch-8       0.00           0.00        ~     (all equal)
MostSpecificHostMatchMixed/1000DestRuleWildcardNoMatch-8      0.00           0.00        ~     (all equal)
MostSpecificHostMatchMixed/5000DestRuleWildcardNoMatch-8      0.00           0.00        ~     (all equal)
MostSpecificHostMatchMixed/10DestRuleWildcardMatch-8          1.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/50DestRuleWildcardMatch-8          1.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/100DestRuleWildcardMatch-8         1.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/1000DestRuleWildcardMatch-8        1.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/5000DestRuleWildcardMatch-8        1.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/10NeedleWildcardNoMatch-8          0.00           0.00        ~     (all equal)
MostSpecificHostMatchMixed/50NeedleWildcardNoMatch-8          0.00           0.00        ~     (all equal)
MostSpecificHostMatchMixed/100NeedleWildcardNoMatch-8         0.00           0.00        ~     (all equal)
MostSpecificHostMatchMixed/1000NeedleWildcardNoMatch-8        0.00           0.00        ~     (all equal)
MostSpecificHostMatchMixed/5000NeedleWildcardNoMatch-8        0.00           0.00        ~     (all equal)
MostSpecificHostMatchMixed/10NeedleWildcardMatch-8            1.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/50NeedleWildcardMatch-8            1.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/100NeedleWildcardMatch-8           1.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/1000NeedleWildcardMatch-8          1.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMixed/5000NeedleWildcardMatch-8          1.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMultiMatch/DestRuleWildcard-8            4.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
MostSpecificHostMatchMultiMatch/NeedleWildcard-8              4.00 ± 0%      0.00        ~     (p=1.000 n=1+1)
```